map fixed
few notes on what needs to be done.

we need to get the search key working so we need to do the geo thing

we need it to zoom in on the pin when it is created but not every time
so in the function that is called by the button press it could modify
the pin position to the new location then zoom to it

we need to find out how we can integrat the directions and location
storage in.